Allow `SwitchUser.switch_back` to be considered even when not calling `#remember_user`.
We need to be able to tell the gem that we want the user to be
remembered so that we can switch back to it. It was unclear how it
should've been done, so this is making sure that we don't have to make
two calls to the `SwitchUser API`, just to make sure the original user
will be remembered.

Related Issue: flyerhzm/switch_user/issues/55